[ADD] website_sale_delivery: Add tour test for free delivery charges over certain amount
This commit adds a tour for the condition where the delivery charges, even though,
was not added to the charges, was shown in the delivery charges aside (even if the payment
was over the amount where free delivery was available). This adds a test that it should be
"0.0" instead. Was tested before the fix and after the fix.